medimeals flutter/backend/requirements.txt
flask
pandas
scikit-learn
requests
Pillow
fpdf
